+++ src/main/java/it/pagopa/pn/external/registries/rest/v1/InfoDomicileController.java
@@ -1,0 +1,70 @@
+package it.pagopa.pn.external.registries.rest;
+
+import it.pagopa.pn.external.registries.api.mock.InfoDomicilieImpl;
+import it.pagopa.pn.external.registries.exceptions.PnInternalException;
+import it.pagopa.pn.external.registries.generated.openapi.server.recipient.domicile.v1.api.InfoDomicileApi;
+import it.pagopa.pn.external.registries.generated.openapi.server.recipient.domicile.v1.dto.AnalogDomicileDto;
+import it.pagopa.pn.external.registries.generated.openapi.server.recipient.domicile.v1.dto.DigitalDomicileDto;
+import it.pagopa.pn.external.registries.generated.openapi.server.recipient.domicile.v1.dto.ProblemDto;
+import it.pagopa.pn.external.registries.generated.openapi.server.recipient.domicile.v1.dto.RecipientTypeDto;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.TypeMismatchException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ServerWebExchange;
+import reactor.core.publisher.Mono;
+
+import javax.validation.ConstraintViolationException;
+import java.util.UUID;
+
+@RestController
+@Slf4j
+public class InfoDomicileController implements InfoDomicileApi {
+    @Override
+    public Mono<ResponseEntity<AnalogDomicileDto>> getOneAnalogDomicile(RecipientTypeDto recipientType, UUID opaqueId, ServerWebExchange exchange) {
+        return InfoDomicilieImpl.getOneAnalogDomicile(recipientType, opaqueId, exchange);
+    }
+
+    @Override
+    public Mono<ResponseEntity<DigitalDomicileDto>> getOneDigitalDomicile(RecipientTypeDto recipientType, UUID opaqueId, ServerWebExchange exchange) {
+        return InfoDomicilieImpl.getOneDigitalDomicile(recipientType, opaqueId, exchange);
+    }
+
+    // catch exception when recipientType = PG (only PF is allowed)
+    @ExceptionHandler({PnInternalException.class})
+    public ResponseEntity<ProblemDto> handleInternalException(PnInternalException ex) {
+        ProblemDto p = new ProblemDto();
+        p.setStatus(HttpStatus.BAD_REQUEST.value());
+        p.setTitle(null);
+        p.setDetail(ex.getMessage());
+        p.setErrors(null);
+
+        return ResponseEntity.badRequest().body(p);
+    }
+
+    // catch invalid opaqueId
+    @ExceptionHandler({ConstraintViolationException.class})
+    public ResponseEntity<ProblemDto> handleValidationException(ConstraintViolationException ex) {
+        ProblemDto p = new ProblemDto();
+        p.setStatus(HttpStatus.BAD_REQUEST.value());
+        p.setTitle("Validation Error");
+        p.setDetail(ex.getMessage());
+        p.setErrors(null);
+
+        return ResponseEntity.badRequest().body(p);
+    }
+
+    // catch invalid recipientType (PF or PG)
+    @ExceptionHandler({TypeMismatchException.class})
+    public ResponseEntity<ProblemDto> handleIllegalArgumentException(TypeMismatchException ex) {
+        ProblemDto p = new ProblemDto();
+        p.setStatus(HttpStatus.BAD_REQUEST.value());
+        p.setTitle("Validation Error");
+        p.setDetail(ex.getRootCause().getMessage());
+        p.setErrors(null);
+
+        return ResponseEntity.badRequest().body(p);
+    }
+}
